requirements.txt
asgiref==3.8.1
black==24.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
